added a new module named 'evaluation'
